- Fixed some mistakes v2
