+++ src/main/java/com/shop/controller/car/DeleteCarController.java
@@ -1,0 +1,23 @@
+package com.shop.controller.car;
+
+import com.shop.lib.Injector;
+import com.shop.service.CarService;
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class DeleteCarController extends HttpServlet {
+    private static final Injector injector = Injector.getInstance("com.shop");
+    private final CarService carService = (CarService) injector
+            .getInstance(CarService.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        Long carId = Long.valueOf(req.getParameter("id"));
+        carService.delete(carId);
+        resp.sendRedirect(req.getContextPath() + "/car/all-cars");
+    }
+}
