+++ src/main/java/com/shop/controller/car/GetAllCarsController.java
@@ -1,23 +1,25 @@
-package com.shop.controller.drivers;
+package com.shop.controller.car;
 
 import com.shop.lib.Injector;
-import com.shop.service.DriverService;
+import com.shop.model.Car;
+import com.shop.service.CarService;
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class DeleteDriversController extends HttpServlet {
+public class GetAllCarsController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.shop");
-    private final DriverService driverService = (DriverService) injector
-            .getInstance(DriverService.class);
+    private final CarService carService = (CarService) injector
+            .getInstance(CarService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long driverId = Long.valueOf(req.getParameter("id"));
-        driverService.delete(driverId);
-        resp.sendRedirect(req.getContextPath() + "/drivers/all-drivers");
+        List<Car> cars = carService.getAll();
+        req.setAttribute("cars", cars);
+        req.getRequestDispatcher("/WEB-INF/views/car/all-cars.jsp").forward(req, resp);
     }
 }
