+++ src/main/java/com/shop/controller/driver/CreateDriverController.java
@@ -1,16 +1,15 @@
-package com.shop.controller.drivers;
+package com.shop.controller.driver;
 
 import com.shop.lib.Injector;
 import com.shop.model.Driver;
 import com.shop.service.DriverService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class GetAllDriversController extends HttpServlet {
+public class CreateDriverController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.shop");
     private final DriverService driverService = (DriverService) injector
             .getInstance(DriverService.class);
@@ -18,8 +17,16 @@ public class GetAllDriversController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Driver> drivers = driverService.getAll();
-        req.setAttribute("drivers", drivers);
-        req.getRequestDispatcher("/WEB-INF/views/drivers/all-drivers.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/driver/create-drivers.jsp").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+            throws IOException {
+        String name = req.getParameter("name");
+        String licenseNumber = req.getParameter("licenseNumber");
+        Driver driver = new Driver(name, licenseNumber);
+        driverService.create(driver);
+        resp.sendRedirect(req.getContextPath() + "all-drivers");
     }
 }
