+++ src/main/java/com/shop/controller/driver/DeleteDriversController.java
@@ -1,23 +1,23 @@
-package com.shop.controller.cars;
+package com.shop.controller.driver;
 
 import com.shop.lib.Injector;
-import com.shop.service.CarService;
+import com.shop.service.DriverService;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class DeleteCarController extends HttpServlet {
+public class DeleteDriversController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.shop");
-    private final CarService carService = (CarService) injector
-            .getInstance(CarService.class);
+    private final DriverService driverService = (DriverService) injector
+            .getInstance(DriverService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long carId = Long.valueOf(req.getParameter("id"));
-        carService.delete(carId);
-        resp.sendRedirect(req.getContextPath() + "/cars/all-cars");
+        Long driverId = Long.valueOf(req.getParameter("id"));
+        driverService.delete(driverId);
+        resp.sendRedirect(req.getContextPath() + "/driver/all-drivers");
     }
 }
