+++ src/main/java/com/shop/controller/driver/GetAllDriversController.java
@@ -1,8 +1,8 @@
-package com.shop.controller.cars;
+package com.shop.controller.driver;
 
 import com.shop.lib.Injector;
-import com.shop.model.Car;
-import com.shop.service.CarService;
+import com.shop.model.Driver;
+import com.shop.service.DriverService;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
@@ -10,16 +10,16 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class GetAllCarsController extends HttpServlet {
+public class GetAllDriversController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.shop");
-    private final CarService carService = (CarService) injector
-            .getInstance(CarService.class);
+    private final DriverService driverService = (DriverService) injector
+            .getInstance(DriverService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Car> cars = carService.getAll();
-        req.setAttribute("cars", cars);
-        req.getRequestDispatcher("/WEB-INF/views/cars/all-cars.jsp").forward(req, resp);
+        List<Driver> drivers = driverService.getAll();
+        req.setAttribute("drivers", drivers);
+        req.getRequestDispatcher("/WEB-INF/views/driver/all-drivers.jsp").forward(req, resp);
     }
 }
