--- src/main/java/com/shop/controller/drivers/CreateDriverController.java
@@ -1,32 +1,0 @@
-package com.shop.controller.drivers;
-
-import com.shop.lib.Injector;
-import com.shop.model.Driver;
-import com.shop.service.DriverService;
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class CreateDriverController extends HttpServlet {
-    private static final Injector injector = Injector.getInstance("com.shop");
-    private final DriverService driverService = (DriverService) injector
-            .getInstance(DriverService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/views/drivers/create-drivers.jsp").forward(req, resp);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
-            throws IOException {
-        String name = req.getParameter("name");
-        String licenseNumber = req.getParameter("licenseNumber");
-        Driver driver = new Driver(name, licenseNumber);
-        driverService.create(driver);
-        resp.sendRedirect(req.getContextPath() + "all-drivers");
-    }
-}
