+++ src/main/java/com/shop/controller/manufacturer/CreateManufacturerController.java
@@ -1,16 +1,15 @@
-package com.shop.controller.manufacturers;
+package com.shop.controller.manufacturer;
 
 import com.shop.lib.Injector;
 import com.shop.model.Manufacturer;
 import com.shop.service.ManufacturerService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class GetAllManufacturerController extends HttpServlet {
+public class CreateManufacturerController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("com.shop");
     private final ManufacturerService manufacturerService = (ManufacturerService) injector
             .getInstance(ManufacturerService.class);
@@ -18,9 +17,17 @@ public class GetAllManufacturerController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Manufacturer> manufacturerList = manufacturerService.getAll();
-        req.setAttribute("manufacturers", manufacturerList);
-        req.getRequestDispatcher("/WEB-INF/views/manufacturers/all-manufacturers.jsp")
+        req.getRequestDispatcher("/WEB-INF/views/manufacturer/create-cars.jsp")
                 .forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+            throws IOException {
+        String name = req.getParameter("name");
+        String country = req.getParameter("country");
+        Manufacturer manufacturer = new Manufacturer(name, country);
+        manufacturerService.create(manufacturer);
+        resp.sendRedirect(req.getContextPath() + "all-cars");
     }
 }
