+++ src/main/java/com/shop/exception/DataProcessingException.java
@@ -1,4 +1,4 @@
-package com.shop.exceptions;
+package com.shop.exception;
 
 public class DataProcessingException extends RuntimeException {
     public DataProcessingException(String message, Throwable cause) {
