+++ src/main/webapp/WEB-INF/views/car/driver/add-drivers-to-car.jsp
@@ -6,7 +6,7 @@
 <body>
 <h1>Please enter data car and driver</h1>
 
-<form method="post" action="${pageContext.request.contextPath}/cars/drivers/add-drivers-to-car">
+<form method="post" action="${pageContext.request.contextPath}/car/driver/add-drivers-to-car">
     Please provide car id<input type="number" name="carId">
     Please provide driver id<input type="number" name="driverId">
