+++ src/main/webapp/WEB-INF/views/driver/all-drivers.jsp
@@ -2,28 +2,30 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
-    <title>All manufacturers</title>
+    <title>All drivers</title>
 </head>
 <body>
+<h1>All drivers page</h1>
+
 <table border="1">
     <tr>
         <th>ID</th>
         <th>Name</th>
-        <th>Country</th>
+        <th>License Number</th>
     </tr>
-    <c:forEach var="manufacturer" items="${manufacturers}">
+    <c:forEach var="driver" items="${drivers}">
         <tr>
             <td>
-                <c:out value="${manufacturer.id}"/>
+                <c:out value="${driver.id}"/>
             </td>
             <td>
-                <c:out value="${manufacturer.name}"/>
+                <c:out value="${driver.name}"/>
             </td>
             <td>
-                <c:out value="${manufacturer.country}"/>
+                <c:out value="${driver.licenseNumber}"/>
             </td>
             <td>
-                <a href="${pageContext.request.contextPath}/manufacturers/delete?id=${manufacturer.id}">Delete</a>
+                <a href="${pageContext.request.contextPath}/driver/delete?id=${driver.id}">Delete</a>
             </td>
         </tr>
     </c:forEach>
