+++ src/main/webapp/WEB-INF/views/manufacturer/create-manufacturers.jsp
@@ -6,7 +6,7 @@
 <body>
 <h1>Please enter data about manufacturer</h1>
 
-<form method="post" action="${pageContext.request.contextPath}/manufacturers/create-manufacturers">
+<form method="post" action="${pageContext.request.contextPath}/manufacturer/create-manufacturers">
     Please provide manufacturer name<input type="text" name="name">
     Please provide manufacturer country<input type="text" name="country">
