+++ src/main/webapp/web.xml
@@ -15,101 +15,101 @@
 
     <servlet>
         <servlet-name>allDrivers</servlet-name>
-        <servlet-class>com.shop.controller.drivers.GetAllDriversController</servlet-class>
+        <servlet-class>com.shop.controller.driver.GetAllDriversController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>allDrivers</servlet-name>
-        <url-pattern>/drivers/all-drivers</url-pattern>
+        <url-pattern>/driver/all-drivers</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>createManufacturer</servlet-name>
-        <servlet-class>com.shop.controller.manufacturers.CreateManufacturerController</servlet-class>
+        <servlet-class>com.shop.controller.manufacturer.CreateManufacturerController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>createManufacturer</servlet-name>
-        <url-pattern>/manufacturers/create-manufacturers</url-pattern>
+        <url-pattern>/manufacturer/create-manufacturers</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>createDriver</servlet-name>
-        <servlet-class>com.shop.controller.drivers.CreateDriverController</servlet-class>
+        <servlet-class>com.shop.controller.driver.CreateDriverController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>createDriver</servlet-name>
-        <url-pattern>/drivers/create-drivers</url-pattern>
+        <url-pattern>/driver/create-drivers</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>createCar</servlet-name>
-        <servlet-class>com.shop.controller.cars.CreateCarController</servlet-class>
+        <servlet-class>com.shop.controller.car.CreateCarController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>createCar</servlet-name>
-        <url-pattern>/cars/create-cars</url-pattern>
+        <url-pattern>/car/create-cars</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>addDriverToCar</servlet-name>
-        <servlet-class>com.shop.controller.cars.drivers.AddDriverToCarController</servlet-class>
+        <servlet-class>com.shop.controller.car.driver.AddDriverToCarController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>addDriverToCar</servlet-name>
-        <url-pattern>/cars/drivers/add-drivers-to-car</url-pattern>
+        <url-pattern>/car/driver/add-drivers-to-car</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>allManufacturers</servlet-name>
-        <servlet-class>com.shop.controller.manufacturers.GetAllManufacturerController</servlet-class>
+        <servlet-class>com.shop.controller.manufacturer.GetAllManufacturerController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>allManufacturers</servlet-name>
-        <url-pattern>/manufacturers/all-manufacturers</url-pattern>
+        <url-pattern>/manufacturer/all-manufacturers</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>allCars</servlet-name>
-        <servlet-class>com.shop.controller.cars.GetAllCarsController</servlet-class>
+        <servlet-class>com.shop.controller.car.GetAllCarsController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>allCars</servlet-name>
-        <url-pattern>/cars/all-cars</url-pattern>
+        <url-pattern>/car/all-cars</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>deleteManufacturer</servlet-name>
-        <servlet-class>com.shop.controller.manufacturers.DeleteManufacturerController</servlet-class>
+        <servlet-class>com.shop.controller.manufacturer.DeleteManufacturerController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>deleteManufacturer</servlet-name>
-        <url-pattern>/manufacturers/delete</url-pattern>
+        <url-pattern>/manufacturer/delete</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>deleteDriver</servlet-name>
-        <servlet-class>com.shop.controller.drivers.DeleteDriversController</servlet-class>
+        <servlet-class>com.shop.controller.driver.DeleteDriversController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>deleteDriver</servlet-name>
-        <url-pattern>/drivers/delete</url-pattern>
+        <url-pattern>/driver/delete</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>deleteCar</servlet-name>
-        <servlet-class>com.shop.controller.cars.DeleteCarController</servlet-class>
+        <servlet-class>com.shop.controller.car.DeleteCarController</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>deleteCar</servlet-name>
-        <url-pattern>/cars/delete</url-pattern>
+        <url-pattern>/car/delete</url-pattern>
     </servlet-mapping>
 </web-app>
